+++ lib/reverse_markdown/converters/text.rb
@@ -9,6 +9,8 @@ module ReverseMarkdown
         end
       end
 
+      private
+
       def treat_empty(node)
         parent = node.parent.name.to_sym
         if [:ol, :ul].include?(parent)  # Otherwise the identation is broken
@@ -21,7 +23,17 @@ module ReverseMarkdown
       end
 
       def treat_text(node)
-        escape_keychars node.text.tr("\n\t", '').squeeze(' ')
+        text = remove_border_newlines(node.text)
+        text = remove_inner_newlines(text)
+        escape_keychars text
+      end
+
+      def remove_border_newlines(text)
+        text.gsub(/\A\n+/, '').gsub(/\n+\z/, '')
+      end
+
+      def remove_inner_newlines(text)
+        text.tr("\n\t", ' ').squeeze(' ')
       end
     end